requirements.txt
PyQt6>=6.4.0
pygame>=2.5.0
numpy>=1.24.0
pyttsx3>=2.90
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
Pillow>=10.0.0

# 3D rendering dependencies
ursina>=6.0.0
panda3d>=1.10.13
